Heartbeat of the broker seems to work

+++ stan_runner/nats_message_broker.py
@@ -1,30 +1,31 @@
 from __future__ import annotations
 
 import asyncio
+import datetime
 import json
 import pickle
 import signal
-import traceback
-from pathlib import Path
-
-from nats import NATS
-from nats.js import JetStreamContext
-from nats.js.api import RawStreamMsg
-from nats.aio.msg import Msg
-import datetime
+import socket
 import time
-import humanize
-from nats.js.errors import NotFoundError
 import uuid
 from collections import deque
-from .nats_utils import connect_to_nats
 
-from .cmdstan_runner import CmdStanRunner
+import humanize
+import nats
+import netifaces
+from adams_text_utils import format_txt_list
+from nats.aio.msg import Msg
+from nats.js import JetStreamContext
+from nats.js.api import RawStreamMsg
+from nats.js.errors import NotFoundError
+
 from .ifaces import StanOutputScope, StanResultEngine
-from .nats_utils import name_topic_datadef, name_topic_modeldef, name_topic_run, STREAM_NAME
+from .nats_utils import STREAM_NAME, create_stream
+from .nats_utils import connect_to_nats
 from .worker_capacity_info import WorkerCapacityInfo
 
 WORKER_TIMEOUT_SECONDS = 60
+
 
 # This is a message broker.
 # It listens to the NATS messages with "stan.rundef.<run_id>" subject.
@@ -45,6 +46,87 @@ WORKER_TIMEOUT_SECONDS = 60
 #
 # In future, broker will also learn how to predict the run time and memory requirements of the tasks. It will use this
 # information to decide which worker to send the task to.
+
+class BrokerInfo:
+    _broker_id: str
+    _last_seen: float
+    _hostname: str
+    _network_addresses: dict[str, list[str]]  # All the network interfaces and their addresses
+
+    @staticmethod
+    def CreateFromLocalHost():
+        last_seen = time.time()
+        hostname = socket.gethostname()
+        broker_id = "broker_" + str(uuid.uuid4())[0:8]
+
+        network_addresses = {}
+        for iface in netifaces.interfaces():
+            iface_addresses = netifaces.ifaddresses(iface)
+            if netifaces.AF_INET in iface_addresses:
+                for addr in iface_addresses[netifaces.AF_INET]:
+                    if addr["addr"] == "127.0.0.1":
+                        continue
+                    if addr["addr"][0:4] == "172.":
+                        continue
+                    if iface not in network_addresses:
+                        network_addresses[iface] = []
+                    network_addresses[iface].append(addr["addr"])
+
+        if "default" in netifaces.gateways():
+            default_gateway = netifaces.gateways()["default"]
+            if netifaces.AF_INET in default_gateway:
+                gateway_ip, iface = default_gateway[netifaces.AF_INET]
+                default_entry = network_addresses[iface]
+                # Put the default gateway first
+                del network_addresses[iface]
+                network_addresses = {iface: default_entry, **network_addresses}
+
+        return BrokerInfo(broker_id, last_seen, hostname, network_addresses)
+
+    @staticmethod
+    def CreateFromSerialized(serialized: bytes):
+        d = json.loads(serialized)
+        return BrokerInfo(**d)
+
+    def __init__(self, broker_id: str, last_seen: float, hostname: str, network_addresses: dict[str, list[str]]):
+        self._broker_id = broker_id
+        self._last_seen = last_seen
+        self._hostname = hostname
+        self._network_addresses = network_addresses
+
+    @property
+    def broker_id(self) -> str:
+        return self._broker_id
+
+    @property
+    def last_seen(self) -> float:
+        return self._last_seen
+
+    def update_last_seen(self):
+        self._last_seen = time.time()
+
+    def pretty_print(self):
+        ans = f"""Broker {self._broker_id} \"{self._hostname}\", last seen {humanize.naturaltime(datetime.datetime.fromtimestamp(time.time()) - datetime.datetime.fromtimestamp(self._last_seen))}, with network addresses:\n\n"""
+
+        ifaces = []
+        for iface, addresses in self._network_addresses.items():
+            ifaces.append(f"* {iface}: {format_txt_list(addresses, max_length=5)}")
+
+        return ans + "\n".join(ifaces)
+
+    def serialize(self) -> bytes:
+        d = {
+            "broker_id": self._broker_id,
+            "last_seen": self._last_seen,
+            "hostname": self._hostname,
+            "network_addresses": self._network_addresses
+        }
+
+        return json.dumps(d).encode()
+
+    def __repr__(self):
+        return self.pretty_print()
+
 
 class ModelInfo:
     """Contains lazy methods for accessing the model's data, when required"""
@@ -117,28 +199,38 @@ class WorkerInfo:
 
     def update_last_seen(self):
         self._last_seen = time.time()
+
     def __repr__(self):
         return self.pretty_print()
 
 
 class MessageBroker:
     _server_context: JetStreamContext
-    _server_raw: NATS
+    _server_raw: nats.NATS
 
     _workers: dict[str, WorkerInfo]
     _tasks: dict[str, TaskInfo]
     _task_queue: deque[str]
 
-    _broker_id: str
+    _subscriptions: list | None
 
-    def __init__(self, server_url: str, user: str, password: str):
-        self._server_raw = connect_to_nats(nats_connection=server_url, user=user, password=password)
-        self._server_context = self._server_raw.jetstream()
+    _broker_id: BrokerInfo
+
+    @staticmethod
+    async def Create(server_url: str, user: str, password: str = None):
+        nc = await connect_to_nats(nats_connection=server_url, user=user, password=password)
+        js, stream = await create_stream(nc, permanent_storage=True, stream_name=STREAM_NAME)
+        return MessageBroker(nc, js)
+
+    def __init__(self, nc: nats.NATS, js: JetStreamContext):
+        self._server_raw = nc
+        self._server_context = js
 
         self._workers = {}
         self._tasks = {}
-        self._task_queue = []
-        self._broker_id = "broker_" + str(uuid.uuid4())[0:8]
+        self._task_queue = deque()
+        self._broker_id = BrokerInfo.CreateFromLocalHost()
+        print(self._broker_id.pretty_print())
 
     def prune_workers(self):
         """Removes workers, that have not advertised themselves in the last WORKER_TIMEOUT_SECONDS seconds"""
@@ -150,7 +242,7 @@ class MessageBroker:
         for worker_id in to_delete:
             del self._workers[worker_id]
 
-    async def handle_worker_adv(self, message:Msg):
+    async def handle_worker_adv(self, message: Msg):
         """Called when a worker advertises itself"""
         worker_id = message.subject.split(".")[-1]
         worker_props = pickle.loads(message.data)
@@ -176,7 +268,6 @@ class MessageBroker:
         self._tasks[task_hash] = task
         self._task_queue.append(task_hash)
         await self.try_sending_tasks_to_workers()
-
 
     async def try_sending_tasks_to_workers(self):
         """Checks if there is an opportunity to send tasks to workers"""
@@ -207,45 +298,72 @@ class MessageBroker:
 
     async def keep_alive(self):
         while True:
-            last_message: RawStreamMsg = await self._server_context.get_last_msg(stream_name=STREAM_NAME,
-                                                                                 subject="stan.broker.alive")
-            if last_message is not None:
-                if last_message.headers["broker_id"] != self._broker_id:
-                    raise Exception(f"Broker id mismatch: {last_message.headers['broker_id']} != {self._broker_id}")
+            print(f"Broker {self._broker_id} is alive")
+            try:
+                last_message: RawStreamMsg | None = await self._server_context.get_last_msg(stream_name=STREAM_NAME,
+                                                                                            subject="stan.broker.alive")
+            except nats.js.errors.NotFoundError:
+                last_message = None
 
-                time_to_last_message = time.time() - last_message.headers["timestamp"]
+            if last_message is not None:
+                time_to_last_message = time.time() - float(last_message.headers["timestamp"])
+                if last_message.headers["broker_id"] != self._broker_id.broker_id:
+                    if time_to_last_message < WORKER_TIMEOUT_SECONDS:
+                        raise Exception(f"Broker id mismatch: {last_message.headers['broker_id']} != {self._broker_id}")
             else:
                 time_to_last_message = float("inf")
 
-            if (time_to_wait := min(0, WORKER_TIMEOUT_SECONDS - time_to_last_message)) > 0:
+            if (time_to_wait := max(0, WORKER_TIMEOUT_SECONDS - time_to_last_message)) > 0:
                 await asyncio.sleep(time_to_wait)
 
-            await self._server_context.publish("stan.broker.alive", b"1",
-                                               headers={"broker_id": self._broker_id, "timestamp": time.time()})
+            broker_bin = self._broker_id.serialize()
+            if self._subscriptions is None:
+                break
+            await self._server_context.publish("stan.broker.alive", payload=broker_bin, stream=STREAM_NAME,
+                                               headers={"broker_id": self._broker_id.broker_id,
+                                                        "format": "json",
+                                                        "timestamp": str(float(time.time()))})
             if last_message is not None:
                 await self._server_context.delete_msg(STREAM_NAME, last_message.seq)
 
-    def the_loop(self):
+    async def the_loop(self):
+        # Attach the shutdown coroutine to SIGINT and SIGTERM
+        loop = asyncio.get_running_loop()
+        loop.add_signal_handler(signal.SIGINT, lambda: asyncio.create_task(self.shutdown(loop)))
+        loop.add_signal_handler(signal.SIGTERM, lambda: asyncio.create_task(self.shutdown(loop)))
 
-        async def shutdown(signal, loop):
-            print(f"Received exit signal {signal.name}...")
-            tasks = [t for t in asyncio.all_tasks() if t is not asyncio.current_task()]
+        sub1 = await self._server_raw.subscribe("stan.worker_adv.>", cb=self.handle_worker_adv)
+        sub2 = await self._server_raw.subscribe("stan.taskdef.>", cb=self.handle_task)
 
-            [task.cancel() for task in tasks]
+        self._subscriptions = [sub1, sub2]
 
-            print(f"Cancelling {len(tasks)} tasks")
-            await asyncio.gather(*tasks, return_exceptions=True)
-            loop.stop()
+        await self.keep_alive()
 
-        # Attach the shutdown coroutine to SIGINT
-        loop = asyncio.get_event_loop()
-        loop.create_task(self._server_raw.subscribe("stan.worker_adv.>", cb=self.handle_worker_adv))
-        loop.create_task(self._server_raw.subscribe("stan.taskdef.>", cb=self.handle_task))
-        loop.create_task(self.keep_alive())
-        loop.add_signal_handler(signal.SIGINT, lambda: asyncio.create_task(shutdown(signal.SIGINT, loop)))
+    async def shutdown(self, loop):
+        print("Unsubscribing...")
+        for s in self._subscriptions:
+            await s.unsubscribe()
+
+        self._subscriptions = None
 
         try:
-            loop.run_forever()
-        finally:
-            loop.close()
-            print("Successfully shutdown the loop")
+            last_message: RawStreamMsg | None = await self._server_context.get_last_msg(stream_name=STREAM_NAME,
+                                                                                        subject="stan.broker.alive")
+        except nats.js.errors.NotFoundError:
+            last_message = None
+
+        await asyncio.sleep(0)  # Yield to other tasks to finish
+
+        print("Removing broadcast message...")
+        if last_message is not None:
+            await self._server_context.delete_msg(STREAM_NAME, last_message.seq)
+
+
+        print("Received exit signal, shutting down...")
+        tasks = [t for t in asyncio.all_tasks() if t is not asyncio.current_task()]
+
+        print(f"Cancelling {len(tasks)} tasks")
+        [task.cancel() for task in tasks]
+
+        await asyncio.gather(*tasks, return_exceptions=True)
+        loop.stop()
